// src/types/moderatorTypes.ts
export interface ModeratorsSlice {
  isModeratorAuthenticated: boolean;
  moderatorName: string | null;
  moderatorRefreshToken: string | null;
  moderatorAccessToken: string | null;
  moderatorId: string | null;
  error: string | null;
  isLoading: "idle" | "pending";
}

export interface ModeratorLoginInput {
  email: string;
  password: string;
}

export interface ModeratorLoginResponse {
  data: T;
  status: number;
}

export interface ModeratorData {
  moderatorrName: string;
  moderatorId: string;
  accessToken: string;
  refreshToken: string;
  message: string;
}